+++ commodore/cli.py
@@ -6,12 +6,13 @@ from commodore import __git_version__
 from .config import Config
 from .helpers import clean_working_tree
 from .compile import compile as _compile
-from .component_template import ComponentFactory
+from .component.template import ComponentFactory
+from .component.compile import compile_component
 
 pass_config = click.make_pass_decorator(Config)
 
 verbosity = click.option('-v', '--verbose', count=True,
-                         help='Control verbosity. Can be repeated for more verbose output')
+                         help='Control verbosity. Can be repeated for more verbose output.')
 
 
 def _version():
@@ -22,58 +23,79 @@ def _version():
 
 
 @click.group()
-@click.option('--api-url', metavar='URL', help='Lieutenant API URL')
-@click.option('--api-token', metavar='TOKEN', help='Lieutenant API token')
-@click.option('--global-git-base', metavar='URL',
-              help='Base directory for global Git config repositories')
-@verbosity
 @click.version_option(_version(), prog_name='commodore')
+@verbosity
 @click.pass_context
-# pylint: disable=too-many-arguments
-def commodore(ctx, api_url, api_token, global_git_base, verbose):
-    ctx.obj = Config(api_url, api_token, global_git_base, verbose)
+def commodore(ctx, verbose):
+    ctx.obj = Config(verbose=verbose)
 
 
-@commodore.command(short_help='Delete generated files')
+@commodore.group(short_help='Interact with a cluster catalog.')
 @verbosity
 @pass_config
-def clean(config, verbose):
+def catalog(config: Config, verbose):
+    config.update_verbosity(verbose)
+
+
+@catalog.command(short_help='Delete generated files.')
+@verbosity
+@pass_config
+def clean(config: Config, verbose):
     config.update_verbosity(verbose)
     clean_working_tree(config)
 
 
-@commodore.command(short_help='Compile inventory and catalog')
+@catalog.command(name='compile', short_help='Compile the catalog.')
 @click.argument('cluster')
+@click.option('--api-url',
+              envvar='COMMODORE_API_URL',
+              help='Lieutenant API URL.', metavar='URL')
+@click.option('--api-token',
+              envvar='COMMODORE_API_TOKEN',
+              help='Lieutenant API token.', metavar='TOKEN')
+@click.option('--global-git-base',
+              envvar='COMMODORE_GLOBAL_GIT_BASE',
+              help='Base directory for global Git config repositories.', metavar='URL')
 @click.option('--local', is_flag=True, default=False,
-              help=('Run in local mode, Local mode does not try to connect to ' +
-                    'Lieutenant API or fetch/push Git repositories.'))
+              help=('Run in local mode, local mode does not try to connect to ' +
+                    'the Lieutenant API or fetch/push Git repositories.'))
 @click.option('--push', is_flag=True, default=False,
-              help='Push catalog to remote repository. Defaults to False')
+              help='Push catalog to remote repository.')
 @verbosity
 @pass_config
-# pylint: disable=redefined-builtin
-def compile(config, cluster, local, push, verbose):
+# pylint: disable=too-many-arguments
+def compile_catalog(config: Config, cluster, api_url, api_token, global_git_base, local, push, verbose):
     config.update_verbosity(verbose)
+    config.api_url = api_url
+    config.api_token = api_token
+    config.global_git_base = global_git_base
     config.local = local
     config.push = push
     _compile(config, cluster)
 
 
-@commodore.command(short_help='Bootstrap new component')
+@commodore.group(short_help='Interact with components.')
+@verbosity
+@pass_config
+def component(config: Config, verbose):
+    config.update_verbosity(verbose)
+
+
+@component.command(name='new', short_help='Bootstrap a new component.')
 @click.argument('name')
 @click.option('--lib/--no-lib', default=False, show_default=True,
-              help='Add component library template')
+              help='Add a component library template.')
 @click.option('--pp/--no-pp', default=False, show_default=True,
-              help='Add component postprocessing template')
+              help='Add a component postprocessing template.')
 @click.option('--owner', default="projectsyn", show_default=True,
-              help='The GitHub user or project name where the component will be hosted')
+              help='The GitHub user or project name where the component will be hosted.')
 @click.option('--copyright', 'copyright_holder',
               default="VSHN AG <info@vshn.ch>", show_default=True,
-              help='The copyright holder added to the license file')
+              help='The copyright holder added to the license file.')
 @verbosity
 @pass_config
 # pylint: disable=too-many-arguments
-def new_component(config, name, verbose, lib, pp, owner, copyright_holder):
+def component_new(config: Config, name, lib, pp, owner, copyright_holder, verbose):
     config.update_verbosity(verbose)
     f = ComponentFactory(config, name)
     f.library = lib
@@ -81,6 +103,26 @@ def new_component(config, name, verbose, lib, pp, owner, copyright_holder):
     f.github_owner = owner
     f.copyright_holder = copyright_holder
     f.create()
+
+
+@component.command(name='compile', short_help='Compile a single component')
+@click.argument('path', type=click.Path(exists=True, file_okay=False, dir_okay=True))
+@click.option('-f', '--values', multiple=True,
+              type=click.Path(exists=True, file_okay=True, dir_okay=False),
+              help='Specify inventory class in a YAML file (can specify multiple).')
+@click.option('-J', '--search-paths', multiple=True,
+              type=click.Path(exists=True, file_okay=False, dir_okay=True),
+              help='Specify additional search paths.')
+@click.option('-o', '--output',
+              default='./', show_default=True,
+              type=click.Path(file_okay=False, dir_okay=True),
+              help='Specify output path for compiled component.')
+@verbosity
+@pass_config
+# pylint: disable=too-many-arguments
+def component_compile(config: Config, path, values, search_paths, output, verbose):
+    config.update_verbosity(verbose)
+    compile_component(config, path, values, search_paths, output)
 
 
 def main():